windows: Changed the SocketDescriptors vector to long long.

+++ daemon/src/audio/audiortp/audio_zrtp_session.cpp
@@ -59,10 +59,10 @@ AudioZrtpSession::AudioZrtpSession(SIPCall &call, const std::string &zidFilename
           localIP.c_str(), call_.getLocalAudioPort());
 }
 
-std::vector<long>
+std::vector<long long>
 AudioZrtpSession::getSocketDescriptors() const
 {
-    std::vector<long> result;
+    std::vector<long long> result;
     result.push_back(dso->getRecvSocket());
     result.push_back(cso->getRecvSocket());
     return result;
